+++ samples/jquery.widgets/common/helpers.js
@@ -562,5 +562,11 @@ primitives.helpers.controls.getCommonOptionsPanels = function (onUpdate, showDef
     new primitives.helpers.controls.CaptionConfig("In order to achive better greacefull degradation of your diagram use item templates of various sizes instead of CSS scale.", false),
     new primitives.helpers.controls.DropDownBoxConfig("scale", 1.0, "CSS Scale", { "50%": 0.5, "60%": 0.6, "70%": 0.7, "80%": 0.8, "90%": 0.9, "100%": 1.0, "110%": 1.1, "120%": 1.2, "130%": 1.3, "140%": 1.4, "150%": 1.5, "160%": 1.6, "170%": 1.7, "180%": 1.8, "190%": 1.9, "200%": 2.0 }, primitives.helpers.controls.ValueType.Number, onUpdate)
   ]));
+  result.push(new primitives.helpers.controls.PanelConfig("Frame", [
+    new primitives.helpers.controls.CaptionConfig("Displays selected items outside view port area.", false),
+    new primitives.helpers.controls.CheckBoxConfig("showFrame", true, "Show Frame", onUpdate),
+    new primitives.helpers.controls.DropDownBoxConfig("frameInnerPadding", 2, "Frame inner padding", [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], primitives.helpers.controls.ValueType.Thickness, onUpdate),
+    new primitives.helpers.controls.DropDownBoxConfig("frameOuterPadding", 2, "Frame outer padding", [1, 2, 3, 4, 5, 6, 7, 8, 9, 10], primitives.helpers.controls.ValueType.Thickness, onUpdate)
+  ]));
   return result;
 };