+++ src.primitives/Controls/FamDiagram/Tasks/Options/ItemsOptionTask.js
@@ -23,10 +23,12 @@ primitives.famdiagram.ItemsOptionTask = function (optionsTask, defaultItemConfig
     var context = {
       isChanged: false,
       hash: _hash,
-      sourceHash: _sourceHash
+      sourceHash: {}
     };
 
     _data = _dataTemplate.read(_data, optionsTask.getOptions(), "options", context);
+
+    _sourceHash = context.sourceHash["options-items"];
 
     return context.isChanged;
   }
@@ -36,7 +38,7 @@ primitives.famdiagram.ItemsOptionTask = function (optionsTask, defaultItemConfig
   }
 
   function getConfig(itemId) {
-    return _sourceHash["options-items"][itemId];
+    return _sourceHash[itemId];
   }
 
   return {