+++ src.primitives/Controls/OrgDiagram/Tasks/Layout/ApplyLayoutChangesTask.js
@@ -1,36 +1,85 @@
 primitives.orgdiagram.ApplyLayoutChangesTask = function (getGraphics, setLayout, itemsSizesOptionTask,
-  currentControlSizeTask, scaleOptionTask, alignDiagramTask) {
+  currentControlSizeTask, scaleOptionTask, alignDiagramTask, frameSizeTask) {
   var _data = {
-    scrollPanelSize: null
+    viewportSize: null,
+    frameThickness: null,
+    framePlaceholderSize: null
   };
 
   function process() {
     var graphics = getGraphics(),
-      scaleOptions = scaleOptionTask.getOptions(),
+      scale = scaleOptionTask.getOptions().scale,
       itemsSizesOptions = itemsSizesOptionTask.getOptions(),
       contentSize = alignDiagramTask.getContentSize(),
-      scrollPanelSize = currentControlSizeTask.getScrollPanelSize();
+      viewportSize = currentControlSizeTask.getScrollPanelSize(),
+      frameThickness = frameSizeTask.getThickness(),
+      autoSize = (itemsSizesOptions.pageFitMode == primitives.common.PageFitMode.AutoSize);
+
+    /* scaled content size */
+    var scaledContentSize = new primitives.common.Size(contentSize);
+    scaledContentSize.scale(scale);
+
+    if (autoSize) {
+      /* resize element to fit placeholder if control in autosize mode */
+      viewportSize = new primitives.common.Size(scaledContentSize.width + 25, scaledContentSize.height + 25);
+      viewportSize.addThickness(frameThickness);
+      viewportSize.cropBySize(itemsSizesOptions.autoSizeMaximum);
+      viewportSize.maxSize(itemsSizesOptions.autoSizeMinimum);//ignore jslint
+
+      // disable frame if its square space is bigger than viewport
+      var framedViewportSize = new primitives.common.Size(viewportSize);
+      framedViewportSize.removeThickness(frameThickness);
+
+      if(viewportSize.space() < framedViewportSize.space() * 2) {
+        viewportSize = framedViewportSize;
+      } else {
+        frameThickness = new primitives.common.Thickness( 0, 0, 0, 0 );
+      }
+    }
+
+    var controlSize = new primitives.common.Size(viewportSize);
+    controlSize.addThickness(frameThickness);
+
+
+    var framePlaceholderSize = new primitives.common.Size(controlSize);
+    framePlaceholderSize.scale(1 / scale);
 
     graphics.resize("placeholder", contentSize.width, contentSize.height);
+    graphics.resize("framePlaceholder", framePlaceholderSize.width, framePlaceholderSize.height );
 
-    _data.scrollPanelSize = setLayout({
-      scale: (scaleOptions.scale),
+    setLayout({
+      scale: scale,
+      autoSize: autoSize,
       contentSize: contentSize,
-      scrollPanelSize: scrollPanelSize,
-      autoSize: (itemsSizesOptions.pageFitMode == primitives.common.PageFitMode.AutoSize),
-      autoSizeMaximum: (itemsSizesOptions.autoSizeMaximum),
-      autoSizeMinimum: (itemsSizesOptions.autoSizeMinimum)
+      scaledContentSize: scaledContentSize,
+      viewportSize: viewportSize,
+      frameThickness: frameThickness,
+      controlSize: controlSize,
+      framePlaceholderSize: framePlaceholderSize
     });
 
+    _data.viewportSize = viewportSize;
+    _data.frameThickness = frameThickness;
+    _data.framePlaceholderSize = framePlaceholderSize;
     return true;
   }
 
   function getOptimalPanelSize() {
-    return new primitives.common.Size(_data.scrollPanelSize.width - 25, _data.scrollPanelSize.height - 25);
+    return new primitives.common.Size(_data.viewportSize.width - 25, _data.viewportSize.height - 25);
+  }
+
+  function getScrollPanelSize() {
+    return new primitives.common.Size(_data.viewportSize.width, _data.viewportSize.height);
+  }
+
+  function getFrameThickness() {
+    return new primitives.common.Thickness(_data.frameThickness);
   }
 
   return {
     process: process,
-    getOptimalPanelSize: getOptimalPanelSize
+    getOptimalPanelSize: getOptimalPanelSize,
+    getScrollPanelSize: getScrollPanelSize,
+    getFrameThickness: getFrameThickness
   };
 };